+++ tests/translation/fetching/test_sql_fetcher.py
@@ -76,15 +76,3 @@ def connection_string(data):
 def insert_data(engine, data):
     for table, table_data in data.items():
         table_data.to_sql(table, engine, index=False)
-
-
-@pytest.fixture(scope="module")
-def data():
-    return {
-        "animals": pd.DataFrame(
-            {"id": [0, 1, 2], "name": ["Tarzan", "Morris", "Simba"], "is_nice": [False, True, True]}
-        ),
-        "humans": pd.DataFrame({"id": [1991, 1999], "name": ["Richard", "Sofia"], "gender": ["Male", "Female"]}),
-        "big_table": pd.DataFrame({"id": range(100)}),
-        "huge_table": pd.DataFrame({"id": range(1000)}),
-    }