+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- The `Fetcher` interface. The old `Fetcher` class in now called `AbstractFetcher`.
+- Support for fetching using multiple fetchers
 - The `fetching.types` module.
+- The `Fetcher` interface. The old `Fetcher` class in now called `AbstractFetcher`.
 - The `action_level.ActionLevel` enum type.
 
 ### Changed